# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/restaurants": {
            "post": {
                "description": "Create a new restaurant with name, email, phone, logo_url, banner_url,...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant"
                ],
                "summary": "Create restaurant",
                "parameters": [
                    {
                        "description": "Restaurant create payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restaurantapp.CreateRestaurantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create restaurant successfully",
                        "schema": {
                            "$ref": "#/definitions/app.CreateRestaurantSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/restaurants/combobox": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get option/combobox groups and items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant"
                ],
                "summary": "Get combobox of a menu item",
                "responses": {
                    "200": {
                        "description": "Get restaurant combobox successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/app.GetComboboxRestaurantSuccessResponseDoc"
                            }
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
        "/api/restaurants/{id}": {
            "get": {
                "description": "Get detailed information of a restaurant using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant"
                ],
                "summary": "Get restaurant by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get restaurant successfully",
                        "schema": {
                            "$ref": "#/definitions/app.GetRestaurantByIDSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            },
            "put": {
                "description": "Update restaurant fields such as name, address, contact info, logo, banner, etc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant"
                ],
                "summary": "Update restaurant information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Restaurant update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/restaurantapp.UpdateRestaurantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update restaurant successfully",
                        "schema": {
                            "$ref": "#/definitions/app.UpdateRestaurantSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a restaurant and its related data using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant"
                ],
                "summary": "Delete restaurant by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/app.DeleteRestaurantSuccessResponseDoc"
                        }
                    },
                    "default": {
                        "description": "Errors",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorDoc"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.CreateRestaurantSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/restaurantapp.CreateRestaurantResponse"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.DeleteRestaurantSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.GetComboboxRestaurantSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Combobox"
                    }
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.GetRestaurantByIDSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/restaurantapp.GetRestaurantByIDResponse"
                },
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "app.UpdateRestaurantSuccessResponseDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
...
        "response.Combobox": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                },
                "value": {}
            }
        },
        "response.ErrorDoc": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "string"
                }
            }
        },
        "restaurant.DayOfWeek": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6
            ],
            "x-enum-comments": {
                "Friday": "5",
                "Monday": "1",
                "Saturday": "6",
                "Sunday": "0",
                "Thursday": "4",
                "Tuesday": "2",
                "Wednesday": "3"
            },
            "x-enum-descriptions": [
                "0",
                "1",
                "2",
                "3",
                "4",
                "5",
                "6"
            ],
            "x-enum-varnames": [
                "Sunday",
                "Monday",
                "Tuesday",
                "Wednesday",
                "Thursday",
                "Friday",
                "Saturday"
            ]
        },
        "restaurantapp.CreateRestaurantRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "banner_url": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "hours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restaurantapp.RestaurantHoursBase"
                    }
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "website_url": {
                    "type": "string"
                }
            }
        },
        "restaurantapp.CreateRestaurantResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "restaurantapp.GetRestaurantByIDResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "banner_url": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "hours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restaurantapp.RestaurantHoursBase"
                    }
                },
                "is_active": {
                    "type": "boolean"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "website_url": {
                    "type": "string"
                }
            }
        },
        "restaurantapp.RestaurantHoursBase": {
            "type": "object",
            "properties": {
                "close_time": {
                    "type": "string"
                },
                "day": {
                    "$ref": "#/definitions/restaurant.DayOfWeek"
                },
                "open_time": {
                    "type": "string"
                }
            }
        },
        "restaurantapp.UpdateRestaurantRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "banner_url": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "hours": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restaurantapp.RestaurantHoursBase"
                    }
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "website_url": {
                    "type": "string"
                }
            }
        },
...
    }
}